feat: add aspect ratio support by calculating client-side (#201)
## Description

This PR allows the developer to specify a desired aspect ratio (AR) when creating an image. 

This PR is based off the awesome work from @maxkolyanov!

This fixes #161

### API

The developer specifies the AR along with normal imgix params, such as `fit`.

e.g.
```jsx
<Imgix
  src={src}
  sizes="(max-width: 30em) 100vw, (max-width: 50em) 50vw"
  imgixParams={{ ar: "2:1" }}
/>
```

The AR has to be specified in the form `number:number`. When an invalid AR is detected, no height values will be generated, and a warning is thrown.

### TODO

- [N/A] Audit tests
- [x] Round `h` up to comply with API spec 
- [x] Add Max as contributor

### New Feature

- [x] If this is a big feature with breaking changes, consider [opening an issue][issues] to discuss first. This is completely up to you, but please keep in mind that your PR might not be accepted.
- [x] Run unit tests to ensure all existing tests are still passing
- [x] Add new passing unit tests to cover the code introduced by your PR
- [x] Update the readme
- [x] Add some [steps](#steps-to-test) so we can test your cool new feature!
- [x] The PR title is in the [conventional commit](#conventional-commit-spec) format: e.g. `feat(<area>): added new way to do this cool thing #issue-number`
- [x] Add your info to the [contributors](#packagejson-contributors) array in package.json!

## Steps to Test

Review changes to unit tests.