feat: create imgixProvider component
Create a Provider that stores and combines the provider-props and the
chidlren's props.

This commit creates a Provider component, <ImgixProvider />. The
component uses the React.createContext() API to create a props context
that can be modified and passed on to its child components. This is
useful as it allows for HOF that can format and combine imgixProps
before they're passed on <ReactImgix />. In order to, for example,
support 2-part URLs, IE domain + path, we can use the Provider to pass
these props to a HOF props merger, format the 2 parts into one combined
URL and finally pass the combined props to <ReactImgix />.

This allows us to leave the current API for <ReactImgix> unchaned while
also unchanged while also adding support for new props now and in the
future.

For example, the ImgixProvider can store a domain prop that is shared
across multiple <ReactImgix> components, each with a distinc src prop.
Inside the Provider, the props reducer combined the shared domain prop
with the src props to create a one-step URL.