feat: create mergeComponentProps & processProps
- `mergeComponentPropsHOF()`: tries to invoke `useImgixContext()`. If
context is `undefined`, it's being accessed outside of `ImgixContext`
provider and the Component is returned as is. Otherwise, the Provider
and Child Component props are combined and passed down as the Child
Component's props.

- `processPropsHOF()`: formats a Component's props to adhere to imgix
  client's specifications.

This commit also creates and the `mergeProps()`, `formatProps()`,
`formatSrc()` and `hasDomain()` helper functions.

- hasDomain: returns `true` if the image has a domain, `false` if not.
- formatSrc: creates a 1-step URL from `domain` and `src` Strings.
- formatProps: formats the following values in the props Object:
  - `width`: if negative gets set to `null`.
  - `height`: if negative gets set to `null`.
  - `src`: concatenated to `domain` if `src` defined and has no domain.
- mergeProps: Merges the `src` object into the `destination` object.
Destination values are overwritten by source values. Source properties
that resolve to `undefined` are skipped if a destination value exists.
- processPropsHOF: formats a Component's props to adhere to imgix
client's specifications.