feat: wrap ReactImgix with props HOFs
This commit wraps all the ReactImgix components with the higher order
functions that process and merge the Providers props with the Component
props.

This simplifies the usage of the ImgixContext wrapper for the user. As
the HOF will apply the context and merge its props with the user's imgix
Component props behind the scenes. In brief, this makes it so the user
only needs to worry about adding the ImgixContext component.